requirements.txt
imageio
matplotlib
opencv-python
pandas
pillow
python-Levenshtein
scikit-image
scipy
torch
torchaudio
torchvision
tensorboard
tqdm
torchinfo